Calculate square root of integer

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -9,7 +9,7 @@
  */
 int fndSQ(int begin, int end, int v)
 {
-	int center;
+	long center;
 
 	if (end >= begin)
 	{
